+++ src/types.js
@@ -232,3 +232,11 @@ export interface IBaseDatabase {
   insertOne(collectionName: string, ...args: Array<any>): Promise<*>;
   remove(collectionName: string, query: Object): Promise<*>;
 }
+
+
+export interface ILogger {
+  static error(params: Array<any>): void,
+  static info(params: Array<any>): void,
+  static log(params: Array<any>): void,
+  static warn(params: Array<any>): void,
+}
